backend/api_service/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
Markdown==3.4.4
packaging==23.1
Pillow==10.0.0
PyJWT==2.8.0
pypng==0.20220715.0
pytz==2023.3
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
whitenoise==6.5.0
django-cors-headers==4.2.0
requests==2.31.0
django-jazzmin==2.6.0
python-dotenv==1.0.0
reportlab==4.0.8
pandas==2.1.4
xhtml2pdf>=0.2.15
django-ckeditor==6.7.0
packaging==23.1
fcm-django>=2.0.1